0.2.7

+++ LGFSwiftTool/LGFExtension/LGF+Double.swift
@@ -1,5 +1,5 @@
 //
-//  LGF+CGFloat.swift
+//  LGF+Double.swift
 //  LGFSwiftTool
 //
 //  Created by apple on 2019/8/20.
@@ -9,16 +9,16 @@
 #if canImport(UIKit)
 import UIKit
 
-public extension CGFloat {
+public extension Double {
     
     // MARK: - 获取相对于屏幕宽度
     func lgf_ScreenW() -> CGFloat {
-        return self *  UIDevice.lgf_ScreenW / 375.0
+        return CGFloat(self) * UIDevice.lgf_ScreenW / 375.0
     }
     
     // MARK: - 获取相对于屏幕高度
     func lgf_ScreenH() -> CGFloat {
-        return self *  UIDevice.lgf_ScreenH / 667.0
+        return CGFloat(self) * UIDevice.lgf_ScreenH / 667.0
     }
     
 }
